requirements.txt
PySide6
Pillow
python-pptx
